:recycle: Refactor and fixed changed requested
