+++ src/components/Navbar/index.tsx
@@ -3,6 +3,9 @@ import { useNavigate } from 'react-router-dom';
 import './styles.scss';
 const Navbar = () => {
   const navigate = useNavigate();
+  const getStarted = () => {
+    navigate('/signup');
+  };
   return (
     <div className="navbar">
       <a className="navbar-brand">Bookry</a>
@@ -24,7 +27,7 @@ const Navbar = () => {
             </a>
           </li>
           <li>
-            <a className="nav-link-item fourth-child" onClick={() => navigate('/signup')}>
+            <a className="nav-link-item fourth-child" onClick={getStarted}>
               Get Started
             </a>
           </li>
