+++ src/types/button.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from 'react';
+
+export type ButtonProps = {
+  text: string;
+  onPress: () => void;
+  type?: string;
+  icon?: ReactNode;
+  loading?: boolean;
+  disabled?: boolean;
+};
