+++ src/types/index.ts
@@ -1,2 +1,4 @@
 export * from './books';
 export * from './user';
+export * from './button';
+export * from './input';
